So many things, Abed, so many things.

+++ Breadit/reddit/src/main/java/me/williamhester/models/Message.java
@@ -170,7 +170,7 @@ public class Message implements Votable, Parcelable {
     }
 
     @Override
-    public long getScore() {
+    public int getScore() {
         return -1;
     }
